machtia/requirements.txt
Django==4.2
djangorestframework==3.14.0
mongoengine==0.28.0
pymongo==4.6.0
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.0
django-cors-headers==4.3.0
requests==2.31.0
